fix compare of integers

+++ src/main/java/nl/mindef/c2sc/nbs/olsr/pud/uplink/server/distributor/impl/DistributorImpl.java
@@ -326,7 +326,7 @@ public class DistributorImpl extends Thread implements Distributor {
 					}
 
 					if ((myIPAddresses.isMe(clusterLeaderGatewayIp) || myIPAddresses.isMe(clusterLeaderMainIp))
-							&& (clusterLeaderGatewayPort == uplinkUdpPort)) {
+							&& (clusterLeaderGatewayPort.intValue() == uplinkUdpPort.intValue())) {
 						/* do not relay to ourselves */
 						if (logger.isDebugEnabled()) {
 							logger.debug("this is me: skipping");
